fix: move `enforce` to hook definition
With the `enforce` flag at the top-most level, it did not seem like the
HTML transformation would actually be applied before the HTML is
processed.

Looking at the implementation of Vite itself, it seems like the
`enforce` hook should be defined on each hook instead

https://github.com/vitejs/vite/blob/41167277d7c14cbc53877480d0a322bcb1bedd1f/packages/vite/src/node/plugins/html.ts#L403

Moving it there seemed to fix the build!