chore: add test helper to wait for Jest assertion
This is based on my QUnit plugin that has the same purpose

https://github.com/alexlafroscia/qunit-wait-for

This allows us to write an assertion that should become successful at _some point_ in the future; we’re just not sure when. It avoids the pitfall of an explicit timeout before performing the assertion, which can lead to a flakey test.

There is a similar helper that’s part of the `@testing-library/dom` package, which I also tried, but it did not work. I think this is because it’s doing something to look for DOM mutations in order to know when to attempt the assertion again. We just want to poll the assertion until it passes.