feat: expose page path to `context` functions (#42)
* [feature] Added a context id to called functions

We use multipage setup a lot in our game projects. To be able to provide each game page with its own data i've added a pageId parameter to the called context functions. See README.md and tests for the details.

* [feature] Update to use the normalised context path

I've update to use the path instead of a new id.

Since I could not get microbundle to build the newly imported vite function, for now i've made a duplicate.

* Update test & docs to reflect using the path as context identifier

* Fix docs for using context callbacks