chore: avoid bundling parts of Vite
By listed Vite as a peer dependency, we can avoid Microbundle trying to bundle the utility that we use. For some reason, processing the Vite file was failing on some unexpected syntax. Regardless, we don't actually want to bundle that code inside this package, so this change makes things more correct.